error fixed

+++ src/redux/modules/answer.js
@@ -127,6 +127,14 @@ const sendAnswerAX = (question_id, content) => {
     };
     axios(options)
       .then((response) => {
+        if (getCookie("is_login") === false) {
+          swal({
+            title: "로그인 필수!",
+            text: "로그인 후 이용가능해요😊",
+            icon: "info",
+          });
+          return;
+        }
         console.log(response.data);
         // 여기서 delete 를 써서 리덕스 정리 한 번 업데이트 해준다.
         dispatch(deleteQuestion(question_id));
